+++ useCases/manchester/src/main/java/de/tum/bgu/msm/health/PersonHealthMCR.java
@@ -28,7 +28,7 @@ public class PersonHealthMCR implements PersonWithSchool, PersonHealth {
     //for exposure model
     private Map<Mode, Float> weeklyMarginalMetHours = new HashMap<>();
     private float weeklyMarginalMetHoursSport = 0.f;
-    private Map<String, Float> weeklyAccidentRisks = new HashMap<>();
+    private Map<String, Double> weeklyAccidentRisks = new HashMap<>();
     private Map<String, float[]> weeklyExposureByPollutantByHour = new HashMap<>();
     private Map<String, Float> weeklyExposureByPollutantNormalised;
 
@@ -266,12 +266,12 @@ public class PersonHealthMCR implements PersonWithSchool, PersonHealth {
     }
 
     @Override
-    public float getWeeklyAccidentRisk(String type) {
-        return weeklyAccidentRisks.getOrDefault(type, 0.f);
+    public double getWeeklyAccidentRisk(String type) {
+        return weeklyAccidentRisks.getOrDefault(type, 0.0);
     }
 
     @Override
-    public void updateWeeklyAccidentRisks(Map<String, Float> newRisks) {
+    public void updateWeeklyAccidentRisks(Map<String, Double> newRisks) {
         //newRisks.forEach((k, v) -> weeklyAccidentRisks.merge(k, v, (v1, v2) -> v1 + v2 - v1*v2));
         newRisks.forEach((k, v) -> weeklyAccidentRisks.merge(k, v, (v1, v2) -> v1 + v2));
     }
